TST: split up test_concat.py #37243 - more follows up (#37387)
